backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
sqlalchemy
alembic
psycopg2-binary
python-dotenv
